requirements.txt
PyMuPDF
tqdm
tools
onnxruntime-genai-cuda 
transformers 
accelerate